module14 homework

+++ controllers/cards.js
@@ -8,18 +8,25 @@ const getCards = (req, res) => {
 
 const postCard = (req, res) => {
   const { name, link, user } = req.body;
-
   Card.create({ name, link, owner: user._id })
     .then((card) => res.send(card))
     .catch((err) => res.status(500).send({ message: `Server cannot create card: ${err}` }));
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.id)
+  Card.findById(req.params.id)
+    .populate('owner')
     .orFail(() => new Error('Not Found'))
+    .then((card) => {
+      if (!card.owner._id.equals(req.body.user._id)) {
+        return Promise.reject(new Error('Forbidden'));
+      }
+      return Card.findByIdAndRemove(req.params.id);
+    })
     .then((card) => res.send(card))
     .catch((err) => {
       if (err.message === 'Not Found') res.status(404).send({ error: 'card does not exist' });
+      else if (err.message === 'Forbidden') res.status(403).send({ error: 'Forbidden' });
       else res.status(500).send({ message: `Server cannot delete card: ${err}` });
     });
 };
